requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
Flask==3.0.3
fonttools==4.53.0
fsspec==2024.6.0
gunicorn==22.0.0
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sympy==1.12.1
tbb==2021.12.0
torch==2.3.0+cpu
torchvision==0.18.0+cpu
typing_extensions==4.12.1
urllib3==2.2.1
Werkzeug==3.0.3
